change datasets, X and y, to type np.float64

+++ scRNA_exp/AdaBoost_training.py
@@ -22,8 +22,8 @@ if __name__ == "__main__":
     k = 0
     for train_index, validate_index in kf.split(X):
         # print("TRAIN:", train_index, "TEST:", validate_index)
-        X_train, X_validate = X[train_index].astype(np.float64), X[validate_index].astype(np.float64)
-        y_train, y_validate = y[train_index].astype(np.float64), y[validate_index].astype(np.float64)
+        X_train, X_validate = X[train_index], X[validate_index]
+        y_train, y_validate = y[train_index], y[validate_index]
 
         model = AdaBoostClassifier()
         model.fit(X_train, y_train)
